Refactor ChoosePrepaidcards sheet
