+++ cardstack/src/screens/sheets/ChoosePrepaidCardSheet/strings.ts
@@ -1,0 +1,4 @@
+export const strings = {
+  noPrepaidCardErrorMessage: `You don't own a Prepaid card!\nYou can create one at app.cardstack.com`,
+  noPrepaidCardErrorTitle: 'Ooops!',
+};
